scraper/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
psycopg2-binary>=2.9
tqdm>=4.66
flask>=3.0
